Remove boxing warnings

+++ src/main/java/org/dschroeer/avshow/Config.java
@@ -39,8 +39,8 @@ public class Config {
   static public final int CACHE_SIZE = Integer.getInteger("org.dschroeer.cache_size", 5).intValue();
   static public final AvTaskProducerMode TASK_PRODUCER_MODE = AvTaskProducerMode.valueOf(System.getProperty("org.dschroeer.task_producer_mode", AvTaskProducerMode.RANDOM_PICTURE_MATCHING_AUDIO_TRACK.toString()));
   static public final ImageChangeMode IMAGE_CHANGE_MODE = ImageChangeMode.valueOf(System.getProperty("org.dschroeer.image_change_mode", ImageChangeMode.SIMPLE_CUT.toString()));
-  static public final long IMAGE_FADE_RUNNING_TIME = Long.getLong("org.dschroeer.image_fade_running_time", 500);
-  static public final int IMAGE_FADE_REPAINT_INTERVAL = Integer.getInteger("org.dschroeer.image_fade_repaint_interval", 10);
+  static public final long IMAGE_FADE_RUNNING_TIME = Long.getLong("org.dschroeer.image_fade_running_time", 500).longValue();
+  static public final int IMAGE_FADE_REPAINT_INTERVAL = Integer.getInteger("org.dschroeer.image_fade_repaint_interval", 10).intValue();
   static public final String[] LOCAL_PRODUCER_COMMAND = System.getProperty("org.dschroeer.local_producer_command", "").split(",");
   static public final String AUDIO_TYPES = System.getProperty("org.dschroeer.audio_types", "\\( -iname *.flac -o -iname *.mp3 -o -iname *.ogg -o -iname *.ape \\)");
   static public final String AUDIO_FOLDER = System.getProperty("org.dschroeer.audio_folder", "/tmp");
